Extract useFetchMemePage hook and add error handling
